# erp-domain/src/main/java/com/bigbrightpaints/erp/modules/auth/web/AuthResponse.java
package com.bigbrightpaints.erp.modules.auth.web;

public record AuthResponse(
        String tokenType,
        String accessToken,
        String refreshToken,
        long expiresIn,
        String companyCode,
        String displayName
) {}