# openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/companies/{id}:
    put:
      tags:
      - company-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyDto'
    delete:
      tags:
      - company-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/admin/users/{id}:
    put:
      tags:
      - admin-user-controller
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserDto'
...
  /api/v1/companies:
    get:
      tags:
      - company-controller
      operationId: list
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListCompanyDto'
    post:
      tags:
      - company-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseCompanyDto'
  /api/v1/auth/refresh-token:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /api/v1/auth/mfa/disable:
    post:
      tags:
      - mfa-controller
      operationId: disable
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaDisableRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMfaStatusResponse'
  /api/v1/auth/mfa/activate:
    post:
      tags:
      - mfa-controller
      operationId: activate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaActivateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMfaStatusResponse'
  /api/v1/auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      parameters:
      - name: refreshToken
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
  /api/v1/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
        '428':
          description: MFA verification required
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseMfaChallengeResponse'
  /api/v1/admin/users:
    get:
      tags:
      - admin-user-controller
      operationId: list_1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListUserDto'
    post:
      tags:
      - admin-user-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserDto'
...
  /api/v1/accounting/journal-entries:
    get:
      tags:
      - accounting-controller
      operationId: journalEntries
      parameters:
      - name: dealerId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListJournalEntryDto'
    post:
      tags:
      - accounting-controller
      operationId: createJournalEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseJournalEntryDto'
  /api/v1/accounting/accounts:
    get:
      tags:
      - accounting-controller
      operationId: accounts
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseListAccountDto'
    post:
      tags:
      - accounting-controller
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseAccountDto'
...
  /api/v1/admin/users/{id}/unsuspend:
    patch:
      tags:
      - admin-user-controller
      operationId: unsuspend
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
  /api/v1/admin/users/{id}/suspend:
    patch:
      tags:
      - admin-user-controller
      operationId: suspend
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    CompanyRequest:
      required:
      - code
      - name
      - timezone
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 0
          type: string
        code:
          maxLength: 64
          minLength: 0
          type: string
        timezone:
          maxLength: 64
          minLength: 0
          type: string
    ApiResponseCompanyDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/CompanyDto'
        timestamp:
          type: string
          format: date-time
    CompanyDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        timezone:
          type: string
    UpdateUserRequest:
      required:
      - displayName
      type: object
      properties:
        displayName:
          type: string
        companyIds:
          type: array
          items:
            type: integer
            format: int64
        roles:
          type: array
          items:
            type: string
        enabled:
          type: boolean
    ApiResponseUserDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserDto'
        timestamp:
          type: string
          format: date-time
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        enabled:
          type: boolean
        mfaEnabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        companies:
          type: array
          items:
            type: string
...
    RefreshTokenRequest:
      required:
      - companyCode
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
        companyCode:
          type: string
    AuthResponse:
      type: object
      properties:
        tokenType:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        companyCode:
          type: string
        displayName:
          type: string
...
    MfaDisableRequest:
      type: object
      properties:
        code:
          type: string
        recoveryCode:
          type: string
    ApiResponseMfaStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/MfaStatusResponse'
        timestamp:
          type: string
          format: date-time
    ApiResponseMfaChallengeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/MfaChallengeResponse'
        timestamp:
          type: string
          format: date-time
    MfaStatusResponse:
      type: object
      properties:
        enabled:
          type: boolean
    MfaChallengeResponse:
      type: object
      properties:
        required:
          type: boolean
    MfaActivateRequest:
      required:
      - code
      type: object
      properties:
        code:
          pattern: \d{6}
          type: string
    LoginRequest:
      required:
      - companyCode
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        companyCode:
          type: string
        mfaCode:
          type: string
        recoveryCode:
          type: string
    CreateUserRequest:
      required:
      - companyIds
      - displayName
      - email
      - password
      - roles
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
        companyIds:
          type: array
          items:
            type: integer
            format: int64
        roles:
          type: array
          items:
            type: string
    JournalEntryRequest:
      required:
      - entryDate
      - lines
      - referenceNumber
      type: object
      properties:
        referenceNumber:
          type: string
        entryDate:
          type: string
          format: date
        memo:
          type: string
        dealerId:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalLineRequest'
    JournalLineRequest:
      required:
      - accountId
      - credit
      - debit
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
    ApiResponseJournalEntryDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/JournalEntryDto'
        timestamp:
          type: string
          format: date-time
    JournalEntryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        referenceNumber:
          type: string
        entryDate:
          type: string
          format: date
        memo:
          type: string
        status:
          type: string
        dealerId:
          type: integer
          format: int64
        dealerName:
          type: string
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalLineDto'
    JournalLineDto:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        accountCode:
          type: string
        description:
          type: string
        debit:
          type: number
        credit:
          type: number
    AccountRequest:
      required:
      - code
      - name
      - type
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        type:
          type: string
    AccountDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        type:
          type: string
        balance:
          type: number
    ApiResponseAccountDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/AccountDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListCompanyDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/CompanyDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListUserDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListJournalEntryDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryDto'
        timestamp:
          type: string
          format: date-time
    ApiResponseListAccountDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AccountDto'
        timestamp:
          type: string
          format: date-time